Assert That Certain Conditions Are True

Currently, we're reexporting stopifnot() as assert() verbatim here.

assert(..., exprs, local = TRUE)

Arguments

...
Any number of expressions that return logical(1), each of which should evaluate to TRUE on success and FALSE on failure.

exprs
any number of (typically but not necessarily logical) R expressions, which should each evaluate to (a logical vector of all) TRUE. Use either … or exprs, the latter typically an unevaluated expression of the form

{
   expr1
   expr2
   ....
}

local
(only when exprs is used:) indicates the environment in which the expressions should be evaluated; by default the one where stopifnot() has been called from.

See also

Examples

assert(
    is.atomic("example"),
    is.character("example")
)