Does the Argument Contain a ggplot2 Scale?

Does the Argument Contain a ggplot2 Scale?

checkScale(x, scale = c("continuous", "discrete"), aes = c("colour",
  "fill"), null.ok = FALSE)

check_scale(x, scale = c("continuous", "discrete"), aes = c("colour",
  "fill"), null.ok = FALSE)

testScale(x, scale = c("continuous", "discrete"), aes = c("colour",
  "fill"), null.ok = FALSE)

test_scale(x, scale = c("continuous", "discrete"), aes = c("colour",
  "fill"), null.ok = FALSE)

assertScale(x, scale = c("continuous", "discrete"), aes = c("colour",
  "fill"), null.ok = FALSE, .var.name = checkmate::vname(x),
  add = NULL)

assert_scale(x, scale = c("continuous", "discrete"), aes = c("colour",
  "fill"), null.ok = FALSE, .var.name = checkmate::vname(x),
  add = NULL)

expect_scale(x, scale = c("continuous", "discrete"), aes = c("colour",
  "fill"), null.ok = FALSE, info = NULL, label = vname(x))

Arguments

x
Object.

scale
string. Type of scale, either "continuous" or "discrete".

aes
string. Aesthetic mapping, either "colour" or "fill". Note that ggplot2 prefers British spelling, so we're enforcing that convention here.

null.ok
logical(1). If set to TRUE, x may also be NULL.

.var.name
character(1). Name of the checked object to print in assertions.

add
AssertCollection. Extra information to be included in the message for the testthat reporter. See testthat::expect_that().

info
character. Extra information to be included in the message for the testthat reporter. See testthat::expect_that().

label
character(1). Name of the checked object to print in messages.

Value

boolean flag for are/has/is*() functions.

Stop on error for assert*() functions (see stop() or stopifnot() for details).

Examples

library(ggplot2)

colour_c <- scale_colour_gradient(low = "red", high = "blue")
class(colour_c)

#> [1] "ScaleContinuous" "Scale"           "ggproto"         "gg"             

colour_d <- scale_colour_manual(values = c("red", "blue"))
class(colour_d)

#> [1] "ScaleDiscrete" "Scale"         "ggproto"       "gg"           

fill_c <- scale_fill_gradient(low = "red", high = "blue")
class(fill_c)

#> [1] "ScaleContinuous" "Scale"           "ggproto"         "gg"             

fill_d <- scale_fill_manual(values = c("red", "blue"))
class(fill_d)

#> [1] "ScaleDiscrete" "Scale"         "ggproto"       "gg"           

checkScale(x = colour_c, scale = "continuous", aes = "colour")

#> [1] TRUE

checkScale(x = colour_d, scale = "discrete", aes = "colour")

#> [1] TRUE

checkScale(x = fill_c, scale = "continuous", aes = "fill")

#> [1] TRUE

checkScale(x = fill_d, scale = "discrete", aes = "fill")

#> [1] TRUE