Does the Requested Input Exist in the Environment?

Does the Requested Input Exist in the Environment?

isExisting(x, envir = parent.frame(), inherits = TRUE,
  .xname = get_name_in_parent(x))

areExisting(x, ...)

isNonExisting(x, ...)

areNonExisting(x, ...)

Arguments

x
character. Variable names to check in environment.

envir
environment. Environment.

inherits
logical(1). Should the enclosing frames of the environment be searched?

.xname
Not intended to be used directly.

...
Additional arguments.

Value

are*/contains*/has*/is*: logical(1). Boolean flag (TRUE/FALSE). Intended to be used in combination with assert() function.

check*: TRUE on success or character(1) on failure. Intended to be used in combination with validate() function.

assert*: TRUE on success or error (stop()) on failure.

Note

exists() only supports character(1), so we are exporting isExisting() as a convenience function to check multiple variables in a single call.

Examples

suppressWarnings(rm(x, y))
a <- 1L
b <- 2L

## Pass ====
isExisting(c("a", "b"))

#>    a    b 
#> TRUE TRUE 

areNonExisting(c("x", "y"))

#> [1] FALSE

## Fail ====
isExisting(c("x", "y"))

#>     x     y 
#> FALSE FALSE 

areNonExisting(c("a", "b"))

#> [1] TRUE